+++ R/editBindingConstraint.R
@@ -3,10 +3,11 @@
 #' @description 
 #' `r antaresEditObject:::badge_api_ok()`
 #' 
-#' Update an existing binding constraint in an Antares study.
+#' Update an existing binding constraint in an Antares study.  
+#' The key search value of the constraint is the `id` field
 #' 
-#'
 #' @inheritParams createBindingConstraint
+#' @param group "character" group of the constraint, default value : "default group"
 #' @template opts
 #' 
 #' @seealso [createBindingConstraint()] to create new binding constraints, [removeBindingConstraint()] to remove binding constraints.
@@ -36,6 +37,7 @@ editBindingConstraint <- function(name,
                                   filter_year_by_year = NULL,
                                   filter_synthesis = NULL,
                                   coefficients = NULL,
+                                  group = NULL,
                                   opts = antaresRead::simOptions()) {
   assertthat::assert_that(inherits(opts, "simOptions"))
   
@@ -67,10 +69,13 @@ editBindingConstraint <- function(name,
     return(invisible(opts))
   }
   
-  valuesIn <- values
-  # Ini file
+  # valuesIn <- values
+  # check Ini file names constraints
   pathIni <- file.path(opts$inputPath, "bindingconstraints/bindingconstraints.ini")
+  
+  # initial parameter list
   bindingConstraints <- readIniFile(pathIni, stringsAsFactors = FALSE)
+  
   previds <- lapply(bindingConstraints, `[[`, "id")
   previds <- unlist(previds, use.names = FALSE)
   if(!id %in% previds){
@@ -81,6 +86,7 @@ editBindingConstraint <- function(name,
   bc_update_pos <- which(previds %in% id)
   bc_update <- bindingConstraints[[bc_update_pos]]
   
+  # Initial parameters of constraint to edit
   iniParams <- list(
     name = bc_update$name,
     id = bc_update$id,
@@ -89,19 +95,35 @@ editBindingConstraint <- function(name,
     operator = bc_update$operator
   )
   
+  # if(!is.null(name)) iniParams$name <- name
+  # if(!is.null(id)) iniParams$id <- id
   
-  if(!is.null(name)) iniParams$name <- name
-  if(!is.null(id)) iniParams$id <- id
-  if(!is.null(enabled)) iniParams$enabled <- enabled
-  if(!is.null(timeStep)) iniParams$type <- timeStep
-  if(!is.null(operator)) iniParams$operator <- operator
-  if(!is.null(filter_year_by_year)){
-    if(opts$antaresVersion >= 832) iniParams$`filter-year-by-year` <- filter_year_by_year
-  }
-  if(!is.null(filter_synthesis)){
-    if(opts$antaresVersion >= 832) iniParams$`filter-synthesis` <- filter_synthesis
+  # update parameters
+    # name can be different of id
+  if(!is.null(name)) 
+    iniParams$name <- name
+  if(!is.null(enabled)) 
+    iniParams$enabled <- enabled
+  if(!is.null(timeStep)) 
+    iniParams$type <- timeStep
+  if(!is.null(operator)) 
+    iniParams$operator <- operator
+  
+  # Marginal price granularity (v8.3.2)
+  if (opts$antaresVersion >= 832){
+    if(!is.null(filter_year_by_year))
+      iniParams$`filter-year-by-year` <- filter_year_by_year
+    if(!is.null(filter_synthesis))
+      iniParams$`filter-synthesis` <- filter_synthesis
   }
   
+  # v860
+  if(opts$antaresVersion>=860){
+    if(!is.null(group))
+      iniParams$group <- group
+  }
+    
+  # update constraint parameters with new parameters
   bindingConstraints[[bc_update_pos]]$name <- iniParams$name
   bindingConstraints[[bc_update_pos]]$id <- iniParams$id
   bindingConstraints[[bc_update_pos]]$enabled <- iniParams$enabled
@@ -134,15 +156,64 @@ editBindingConstraint <- function(name,
     }
   }
   
-  values <- .valueCheck(values, bindingConstraints[[bc_update_pos]]$type)
+  # write txt files
+    # v860
+  if(opts$antaresVersion>=860 & !is.null(values))
+    values <- .valueCheck860(values, bindingConstraints[[bc_update_pos]]$type)
+  else
+    values <- .valueCheck(values, bindingConstraints[[bc_update_pos]]$type)
   
   # Write Ini
   writeIni(listData = bindingConstraints, pathIni = pathIni, overwrite = TRUE)
   
   # Write values
-  pathValues <- file.path(opts$inputPath, "bindingconstraints", paste0(id, ".txt"))
-  
-  if(!is.null(valuesIn))write.table(x = values, file = pathValues, col.names = FALSE, row.names = FALSE, sep = "\t")
+  # v860
+  if(opts$antaresVersion>=860){
+    if(!identical(values, character(0))){
+      names_order_ts <- c("lt", "gt", "eq")
+      name_file <- paste0(id, "_", names_order_ts, ".txt")
+      
+      up_path <- file.path(opts$inputPath, "bindingconstraints", name_file)
+      
+      lapply(up_path, function(x, df_ts= values, vect_path= up_path){
+        index <- grep(x = up_path, pattern = x)
+        fwrite(x = data.table::as.data.table(df_ts[[index]]), 
+               file = x, 
+               col.names = FALSE, 
+               row.names = FALSE, 
+               sep = "\t")
+      })
+    }
+    
+  }else{
+    pathValues <- file.path(opts$inputPath, 
+                            "bindingconstraints", 
+                            paste0(id, ".txt"))
+    
+    # read to check timestep
+    suppressWarnings(
+      file_r <- fread(pathValues)
+    )
+    
+    # # check nrow Vs timeStep
+    # nrows <- switch(timeStep,
+    #                 hourly = 24*366,
+    #                 daily = 366,
+    #                 weekly = 366,
+    #                 monthly = 12,
+    #                 annual = 1)
+    # 
+    # # check existing values 
+    # if(!is.null(timeStep) & nrow(file_r)>0)
+    #   if(!nrow(file_r) %in% nrows)
+    #     stop("Incorrect number of rows according to the timeStep")
+
+    if(!identical(values, character(0)))
+      write.table(x = values, 
+                  file = pathValues, 
+                  col.names = FALSE, 
+                  row.names = FALSE, sep = "\t")
+  }
   
   # Maj simulation
   suppressWarnings({